GH-1587: Don't Correct TX Offsets After Seek(s)
Resolves https://github.com/spring-projects/spring-kafka/issues/1587

We should not advance the consumer partition if Seek operations have been
performed; skip fixing the offsets if that condition is detected.

Capture the positions after the poll and check during the fix operation.

This is difficult to write a unit test for; tested with a Boot application,
observing the correct behavior via DEBUG logs.

**cherry-pick to 2.5.x**

(cherry picked from commit bd911ad99cffd212224c3401127b8bc349ea3670)