GH-1983: Support copy of producer factory
Fixes https://github.com/spring-projects/spring-kafka/issues/1983

Preserves post processors for of custom factories

Instrumentations attached to ProducerFactory instances are destroyed if the user adds configOverrides.
This behavior is unexpected for the developer, for example it will undo sleuth instrumentation
that is otherwise kept if the new KafkaTemplate(Map) constructor would be used.

* adds copy for all visible factory properties and tests
* moves the copy feature to ProducerFactory
* fixing review remarks
* removes a left over TODO
* implements a generic copy in the KafkaTemplate to avoid breaking changes
* wraps javadoc at col 90
* Clean up for code style

**Cherry-pick to `2.5.x`, `2.6.x` & `2.7.x`**

# Conflicts:
#	spring-kafka/src/main/java/org/springframework/kafka/core/DefaultKafkaProducerFactory.java
#	spring-kafka/src/main/java/org/springframework/kafka/core/ProducerFactory.java
#	spring-kafka/src/test/java/org/springframework/kafka/core/KafkaTemplateTests.java