DLPR - Generate Header Values Lazily
The `this.whichHeaders.contains` check happens after the byte array (in
case of `maybeAddHeader`) or the `RecordHeader` (in `appendOrReplace`)
has already been created. It may be quite an expensive operation, so in
this PR I propose to pass a supplier into these two functions instead.

Change the signature of appendOrReplace according to the review.