GH-2212: Log listener exception in retry topic flow (#2213)
* GH-2212: Log listener ex. in retry topic flow

Resolves https://github.com/spring-projects/spring-kafka/issues/2212

As in Retry Topic flow recovery is successful, the exception thrown by the listener was never being logged.

Now it logs at DEBUG level for intermediate retries and ERROR level when retries are exhausted.

Also some polishing in TimestampedException as it was cluttering the stacktrace with duplicated information from cause.getMessage(). Now it logs the time when the exception occurred.

* Address TimestampedException review suggestions

Add ListenerExceptionLoggingStrategy

Add unit tests

* Address review comments