GH-2195: DefaultErrorHandler Improvements
Resolves https://github.com/spring-projects/spring-kafka/issues/2195

Add an option to avoid seeks after handling exceptions.

Instead, pause the consumer for one `poll()` and use the remaining records as the
result of that poll.

New methods on `CommonErrorHandler` - `handleOne` for record listeners, returning
a boolean to indicate whether the record was recovered and should not be redelivered.

`handlaBatchAndReturnRemaining` for batch listeners, returning either the complete
set or a subset, e.g. when the `DEH` receives a `BatchListenerExecutionFailedException`
and commits a partial batch.

Also includes the classifier refactoring discussed here
https://github.com/spring-projects/spring-kafka/pull/2185#issuecomment-1078259823

The new logic is disabled by default, we can consider enabling it in 3.0 and remove
the deprecations.

* Fix race - do not call `resume()` on the container; the user might have paused after the error.

* Change Since to 2.9.

* Fix typos.

Co-authored-by: Artem Bilan <abilan@vmware.com>

* Remove unnecessary local variable; add docs.

* Polishing - see commit comment for more details

- move the resume logic to after the invokes and don't resume if pending records
- don't check `isPaused()` after empty poll due to errors; always restore the pending records

* Remove unnecessary boolean; fix deprecation warnings and delegating error handlers.

* Emergency stop container if the consumer returns records while paused after an error.

* Fix race in test - prevent consumer thread from changing pausedConsumers while the test thread is calling revoke/assign.

* Remove System.out().

* Add diagnostics to test.

* Fix race in test; wait until next poll after consumer thread pauses the partitions.

* Fix stubbing in emergency stop test.

* Remove unnecessary boolean.

**Cherry-pick to `2.9.x`**