Remove `provided` deps from the target POM
Related to https://github.com/spring-projects/spring-framework/issues/23486

We have this `providedApi 'org.junit.platform:junit-platform-launcher'`
where its `version` is resolved from `mavenBom "org.junit:junit-bom:$junitJupiterVersion"`.
When we generate a Maven POM, this kind of dependency is present as an `optional` by default.
We previously have overridden this to the `provided`, but it comes without a `version`.

According to Spring Framework decision this kind of transitive dependencies are misleading.

* Remove any `provided` deps in Gradle from the target POM altogether