GH-2464: Fix Delivery Attempt Header (Rare)
Resolves https://github.com/spring-projects/spring-kafka/issues/2464

`KafkaBackoffException` should never be classified as retryable; the design
is such that those exceptions go straight to the recoverer, which will re-throw.

If, however, a user configures a super class as retryable (e.g. `RuntimeException`)
then its classification wins.

This causes a new retry state (`FailedRecord`) to be started, causing an incorrect
value of 2 instead of 1 on the first delivery from a retry topic.

**I will backport - conflicts expected**