GH-2505: Use Correct Brokers for Cluster Id
Resolves https://github.com/spring-projects/spring-kafka/issues/2505

When observation is enabled, we use the context's `KafkaAdmin`, if present,
to determine the cluster id.

The context's admin may be configured with different bootstrap servers than
the producer/consumer (e.g. in Spring Cloud Stream).

When looking up the cluster id for observation, use the appropriate bootstrap
servers for the component.